feat(player): 抽出插件 runtime
