+++ packages/player/src/components/ExtensionContext/ext-ctx.ts
@@ -1,54 +1,9 @@
 import type * as TauriHttp from "@tauri-apps/plugin-http";
-import { fromObject, fromSource, removeComments } from "convert-source-map";
 import type { ComponentType } from "react";
-import { SourceMapConsumer, SourceMapGenerator } from "source-map-js";
 import type { db } from "../../dexie.ts";
 import type ExtensionEnv from "../../extension-env.ts";
 import i18n from "../../i18n.ts";
 import type { ExtensionMetaState } from "../../states/extensionsAtoms.ts";
-
-export async function sourceMapOffsetLines(
-	code: string,
-	sourceRoot: string,
-	lineOffset: number,
-): Promise<[string, string]> {
-	const incomingSourceConv = fromSource(code);
-	if (!incomingSourceConv) return [code, ""];
-	const incomingSourceMap = incomingSourceConv.toObject();
-	const consumer = await new SourceMapConsumer(incomingSourceMap);
-	const generator = new SourceMapGenerator({
-		file: incomingSourceMap.file,
-		sourceRoot: sourceRoot,
-	});
-	consumer.eachMapping((m) => {
-		// skip invalid (not-connected) mapping
-		// refs: https://github.com/mozilla/source-map/blob/182f4459415de309667845af2b05716fcf9c59ad/lib/source-map-generator.js#L268-L275
-		if (
-			typeof m.originalLine === "number" &&
-			0 < m.originalLine &&
-			typeof m.originalColumn === "number" &&
-			0 <= m.originalColumn &&
-			m.source
-		) {
-			generator.addMapping({
-				source:
-					m.source &&
-					`${location.origin}/extensions/${sourceRoot}/${m.source.replace(/^(\.*\/)+/, "")}`,
-				name: m.name,
-				original: { line: m.originalLine, column: m.originalColumn },
-				generated: {
-					line: m.generatedLine + lineOffset,
-					column: m.generatedColumn,
-				},
-			});
-		}
-	});
-	const outgoingSourceMap = JSON.parse(generator.toString());
-	if (typeof incomingSourceMap.sourcesContent !== "undefined") {
-		outgoingSourceMap.sourcesContent = incomingSourceMap.sourcesContent;
-	}
-	return [removeComments(code), fromObject(outgoingSourceMap).toComment()];
-}
 
 export class PlayerExtensionContext
 	extends EventTarget
